refactor: déplacer les fichiers de lancement et de déplacement des éléments de jeu dans le dossier Game
- Déplacement des fichiers de lancement du jeu et de déplacement des éléments de jeu dans le dossier `Game`
- Ajustement des appels de fonctions en conséquence
