+++ Entities/fish.h
@@ -6,7 +6,7 @@
 #include <iostream>
 #include <string>
 #include <vector>
-#include "../camera.h"
+#include "../Game/camera.h"
 #include "../Utility/env.h"
 
 class Fish {
