+++ Game/LaunchGameMulti.h
@@ -1,0 +1,20 @@
+//
+// Created by BreizhHardware on 18/12/2024.
+//
+
+#ifndef LAUNCHGAMEMULTI_H
+#define LAUNCHGAMEMULTI_H
+
+#include <iostream>
+#include <string>
+#include <thread>
+#include <algorithm>
+
+#include "../Utility/env.h"
+#include "decors.h"
+#include "../Entities/fish.h"
+#include "../Utility/utility.h"
+
+int pas_la_fontion_main_enfin_ce_nest_pas_la_fontion_principale_du_programme_mais_une_des_fonctions_principale_meme_primordiale_du_projet_denomme_bloubloulespoissons_mais_celle_ci_elle_lance_en_multijoueur(int argc, std::string args);
+
+#endif //LAUNCHGAMEMULTI_H
