+++ Game/LaunchGameSolo.h
@@ -1,0 +1,20 @@
+//
+// Created by BreizhHardware on 18/12/2024.
+//
+
+#ifndef LAUNCGAMESOLO_H
+#define LAUNCGAMESOLO_H
+
+#include <iostream>
+#include <vector>
+#include <algorithm>
+#include <thread>
+#include "../Utility/env.h"
+#include "../Utility/utility.h"
+#include "../Entities/fish.h"
+#include "../Entities/shark.h"
+#include "decors.h"
+
+int pas_la_fontion_main_enfin_ce_nest_pas_la_fontion_principale_du_programme_mais_une_des_fonctions_principale_meme_primordiale_du_projet_denomme_bloubloulespoissons(int argc, char* args[]);
+
+#endif //LAUNCGAMESOLO_H
